psg_test/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
Brotli==1.0.9
certifi==2022.5.18.1
click==8.1.3
dash==2.4.1
dash-bootstrap-components==1.1.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.1.2
Flask-Compress==1.12
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.4
pandas==1.4.2
plotly==5.8.0
psycopg2-binary
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
tenacity==8.0.1
Werkzeug==2.1.2
zipp==3.8.0
